Pricelist module - Refactoring.
