+++ models/pricelist/pricelist_container_pl.py
@@ -196,122 +196,6 @@ class PricelistContainer(models.Model):
 	# create_products_products
 
 
-# --------------------------------------------- Create or update Products Templates - Button ---
-	@api.multi
-	def create_product_templates(self):
-		"""
-		Pricelist 2019
-		Creates Product Templates 
-		Uses Pricelist Products
-		Avoids Product Duplication
-		"""
-		print('\n\n')
-		print('Create Products 2019')
-
-		# Search all Pricelist products
-		products = self.env['openhealth.product.pricelist'].search([],)
-
-		# Count all
-		count = self.env['openhealth.product.pricelist'].search_count([],)
-
-		# Loop 
-		for pro in products:
-			print('\n')
-			print(pro)
-			print(pro.name)  		# Generates Encode Error
-
-			# Count PTs 
-			count = self.env['product.template'].search_count([
-																('name', '=', pro.name),
-																('pl_price_list', '=', '2019'),
-														])
-			#print(count)
-
-			# Avoids PT Duplication
-			if count == 0:
-
-				# Create 
-				print('CREATE !')
-				product_template = self.env['product.template'].create({
-																			'name': 			pro.name,
-																			'sale_ok': 			True,
-																			'purchase_ok': 		False,
-																			'pl_price_list': 	'2019',
-																			'pl_time_stamp': 	pro.time_stamp,
-																			'type': 			pro.x_type,
-																			'pl_name_short': 	pro.name_short,
-																			'pl_prefix': 		pro.prefix,
-																			'pl_idx': 			pro.idx,
-																			'pl_idx_int': 		pro.idx_int,
-																			'pl_family': 		pro.family,
-																			'pl_subfamily':		pro.subfamily,
-																			'pl_treatment': 	pro.treatment,
-																			'pl_zone': 			pro.zone,
-																			'pl_pathology': 	pro.pathology,
-																			'pl_level': 		pro.level,
-																			'pl_sessions': 		pro.sessions,
-																			'pl_time': 			pro.time,
-																			'list_price': 				pro.price,
-																			'pl_price_vip': 			pro.price_vip,
-																			'pl_price_company': 		pro.price_company,
-																			'pl_price_session': 		pro.price_session,
-																			'pl_price_session_next': 	pro.price_session_next,
-																			'pl_price_max': 			pro.price_max,
-
-																			# Only Prods
-																			'pl_manufacturer': 			pro.manufacturer,
-																			'pl_brand': 				pro.brand,
-				})
-				print(product_template)
-
-			else:
-				print('WRITE !!!')
-
-				# Search
-				product_template = self.env['product.template'].search([
-																		('name', '=', pro.name),
-																		('pl_price_list', '=', '2019'),
-																	])
-				print(product_template)
-
-				# Write 
-				product_template.write({
-																			'name': 			pro.name,
-																			'sale_ok': 			True,
-																			'purchase_ok': 		False,
-																			'pl_price_list': 	'2019',
-																			'pl_time_stamp': 	pro.time_stamp,
-																			'type': 			pro.x_type,
-																			'pl_name_short': 	pro.name_short,
-																			'pl_prefix': 		pro.prefix,
-																			'pl_idx': 			pro.idx,
-																			'pl_idx_int': 		pro.idx_int,
-																			'pl_family': 		pro.family,
-																			'pl_subfamily':		pro.subfamily,
-																			'pl_treatment': 	pro.treatment,
-																			'pl_zone': 			pro.zone,
-																			'pl_pathology': 	pro.pathology,
-																			'pl_level': 		pro.level,
-																			'pl_sessions': 		pro.sessions,
-																			'pl_time': 			pro.time,
-																			'list_price': 				pro.price,
-																			'pl_price_vip': 			pro.price_vip,
-																			'pl_price_company': 		pro.price_company,
-																			'pl_price_session': 		pro.price_session,
-																			'pl_price_session_next': 	pro.price_session_next,
-																			'pl_price_max': 			pro.price_max,
-
-																			# Only Prods
-																			'pl_manufacturer': 			pro.manufacturer,
-																			'pl_brand': 				pro.brand,
-																			})
-				#print(product_template)
-
-		print('Created end !')
-
-	# create_products_2019
-
-
 # ---------------------------------------------------------- Update - Button ---
 	@api.multi
 	def update(self):
@@ -322,11 +206,8 @@ class PricelistContainer(models.Model):
 
 		# Search
 		products = self.env['product.template'].search([
-															('pl_price_list', '=', '2019'),
-														],
-															#order='date_begin asc',
-															#limit=1,
-													)
+                    ('pl_price_list', '=', '2019'),
+                    ],)
 		for product in products:
 			product.update()
 	# update
