+++ views/menus/menus_dev.xml
@@ -11,6 +11,9 @@
             <field name="res_model">product.product</field>
             <field name="view_type">form</field>
             <field name="view_mode">tree,form</field>
+
+            <field name="view_id" ref="product_template_list"/> 
+
             <field name="domain">[]</field>
         </record>
